libmp4: always read the tfra data into 64 bits variables
We need to handle the worst case scenario anyway.
It should not take that much more memory and that avoids
casting pointers.

(cherry picked from commit 3e4bec57cfc24826aca3a80730353017156c14e8)

+++ modules/demux/mp4/libmp4.h
@@ -1577,8 +1577,8 @@ typedef struct
     uint8_t i_length_size_of_trun_num;
     uint8_t i_length_size_of_sample_num;
 
-    uint32_t *p_time;
-    uint32_t *p_moof_offset;
+    uint64_t *p_time;
+    uint64_t *p_moof_offset;
     uint8_t *p_traf_number;
     uint8_t *p_trun_number;
     uint8_t *p_sample_number;
